add stubs and prototypes to make it compile
apart from the used amiga API there seem to be some game-related
functions missing as well (see stubs.c).

this is just the beginning, there's tons of warnings to fix,
functions to declare in headers,
and usage of the amiga function replaced with appropriate SDL calls
if needed.

there's probably a lot of breakage due to usage of functions which are
supposed to return pointers, which default to int return value due to
missing declarations.

Makefile added as well.