readd unused function compare()
this function was used from the original (lost) validity_check()
function.
it was removed in the commit:
"remove hardcoded NO_OF_COMMANDS"

readding it since it a part of the missing validity_check() function
reappeared (chunk of code starting at offset 0x2AC18 in the .adf).

additionally it implements the logic to use a short form of a command
for example "E" instead of "EXERCISE"